transfered part of logic from controllers to services

+++ src/main/java/ru/tsystems/internetshop/service/Impl/ClientAddressServiceImpl.java
@@ -3,11 +3,13 @@ package ru.tsystems.internetshop.service.Impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.ui.Model;
 import ru.tsystems.internetshop.dao.ClientAddressDAO;
 import ru.tsystems.internetshop.model.DTO.ClientAddressDTO;
 import ru.tsystems.internetshop.model.DTO.ClientDTO;
 import ru.tsystems.internetshop.model.entity.ClientAddress;
 import ru.tsystems.internetshop.service.ClientAddressService;
+import ru.tsystems.internetshop.service.OrderService;
 import ru.tsystems.internetshop.util.Mapper;
 
 import java.util.ArrayList;
@@ -26,15 +28,8 @@ public class ClientAddressServiceImpl implements ClientAddressService {
     @Autowired
     private Mapper mapper;
 
-    /**
-     * This method saves client address
-     *
-     * @param clientAddressDTO client address
-     */
-    @Override
-    public void saveClientAddress(ClientAddressDTO clientAddressDTO) {
-        clientAddressDAO.create(mapper.convertToEntity(clientAddressDTO));
-    }
+    @Autowired
+    private OrderService orderService;
 
     /**
      * This method updates client address
@@ -42,8 +37,21 @@ public class ClientAddressServiceImpl implements ClientAddressService {
      * @param clientAddressDTO client address
      */
     @Override
-    public void updateClientAddress(ClientAddressDTO clientAddressDTO) {
-        clientAddressDAO.update(mapper.convertToEntity(clientAddressDTO));
+    public Model updateClientAddress(ClientAddressDTO clientAddressDTO, Model model) {
+        ClientDTO clientDTO = clientAddressDTO.getClient();
+
+        if (orderService.getUnfinishedOrdersByClient(clientDTO).size() != 0)
+            model.addAttribute("errorMessage", "Data change is not possible: you have incomplete orders.");
+        else {
+            clientAddressDAO.update(mapper.convertToEntity(clientAddressDTO));
+
+            clientDTO.setAddresses(getAddressesByClient(clientDTO));
+
+            model.addAttribute("errorMessage", "Your data successfully changed");
+            model.addAttribute("client", clientDTO);
+        }
+
+        return model;
     }
 
     /**
@@ -84,7 +92,42 @@ public class ClientAddressServiceImpl implements ClientAddressService {
      * @param clientAddressDTO client address
      */
     @Override
-    public void deleteAddress(ClientAddressDTO clientAddressDTO) {
-        clientAddressDAO.delete(mapper.convertToEntity(clientAddressDTO));
+    public Model deleteClientAddress(ClientAddressDTO clientAddressDTO, Model model) {
+        ClientDTO clientDTO = clientAddressDTO.getClient();
+
+        if (orderService.getUnfinishedOrdersByClient(clientDTO).size() != 0)
+            model.addAttribute("errorMessage", "Data change is not possible: you have incomplete orders.");
+        else {
+            clientAddressDAO.delete(mapper.convertToEntity(clientAddressDTO));
+
+            clientDTO.setAddresses(getAddressesByClient(clientDTO));
+
+            model.addAttribute("successMessage", "Your address successfully deleted");
+            model.addAttribute("client", clientDTO);
+        }
+
+        return model;
+    }
+
+    /**
+     * This method saves client address
+     *
+     * @param clientAddressDTO client address
+     */
+    @Override
+    public Model createAddress(ClientDTO clientDTO, ClientAddressDTO clientAddressDTO, Model model) {
+        clientAddressDTO.setClient(clientDTO);
+        if (orderService.getUnfinishedOrdersByClient(clientDTO).size() != 0)
+            model.addAttribute("errorMessage", "Data change is not possible: you have incomplete orders.");
+        else {
+            clientAddressDAO.create(mapper.convertToEntity(clientAddressDTO));
+
+            clientDTO.setAddresses(getAddressesByClient(clientDTO));
+
+            model.addAttribute("successMessage", "Your data successfully saved");
+            model.addAttribute("client", clientDTO);
+        }
+
+        return model;
     }
 }
